requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Pydantic for validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Google Cloud Firestore
google-cloud-firestore==2.14.0

# Firebase Admin SDK (for Realtime Database + Auth)
firebase-admin==6.4.0

# Environment variables
python-dotenv==1.0.1

# HTTP client (for health checks, external APIs)
httpx==0.26.0

# Logging
python-json-logger==2.0.7
